feat: Return the number of errors as the status code.
This change includes some adjustments to the various Stream methods:
- The `close(channel)` commands now `defer`
- Encountering an error inside `range` no longer returns
- Each stream now closes its respective channel consistently

Previously, encountering an error early in the pipeline would cause
a hang if the final `os.Exit` in `main.go` was removed. This is because
the errors channel does not close until the `Printer` completes. In the
event of an error, the earlier stages were returning and skipping their
`close` steps.

Now, errors are sent to the error channel normally but do not `return`,
and each channel in the pipeline closes itself as it completes, so the
final channel (`Printer`) can close `errors` and end the program.

The `defer close(x)` steps may no longer be necessary, but it is an
easy way to visually mark them as something that should run regardless
of how the function returns.