feat: Handle errors as test failures.
Instead of passing errors into the error channel, the build, transform,
and run steps now pass any errors along with the test data. This allows
the printers to handle errors as they wish. The errors are also passed
back to the main program, so that it can calcuate the correct status
code to return.

One potential issue with this is that errors from the file load step
may get swallowed (since main doesn't print them).