refactor: Rename parsers.Spec to parsers.Paragraph.
refactor: Rename parsers.Spec to parsers.Paragraph.

This is an attempt to reclaim some of the distinction between the data
we have at the parsing stage, and the data we have at the build stage.
"Spec" and "Test", in software, are sometimes interchangeable.

"Paragraph" was suggested as an alternative for what we're currently
calling parsers.Spec, because it essentially represents a group of
semantically-related lines.